services/iam_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
alembic
pydantic[email]
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-multipart